fix(frontend): cartView
